weekly services updates all stocks and services weekly at 16:05 PM ET. Removed edit button from watchlist. Reduce all sector functions into one service.
