+++ client/src/components/Menu/index.js
@@ -1,2 +1,1 @@
 export * from './Drawer_Menu'
-export * from './Filter_Menu'
