+++ client/src/components/WatchTable/index.jsx
@@ -107,30 +107,6 @@ export function WatchTable({ user }) {
         data={state.data}
         options={{pageSize:10}}
         editable={{
-          onRowAdd: newData =>
-            new Promise(resolve => {
-              setTimeout(() => {
-                resolve()
-                setState(prevState => {
-                  const data = [...prevState.data]
-                  data.push(newData)
-                  return { ...prevState, data }
-                })
-              }, 600)
-            }),
-          onRowUpdate: (newData, oldData) =>
-            new Promise(resolve => {
-              setTimeout(() => {
-                resolve()
-                if (oldData) {
-                  setState(prevState => {
-                    const data = [...prevState.data]
-                    data[data.indexOf(oldData)] = newData
-                    return { ...prevState, data }
-                  })
-                }
-              }, 600)
-            }),
           onRowDelete: oldData =>
             new Promise(resolve => {
               setTimeout(() => {
